Movement done
Spin is commented right now. Off screen buffer is 2x the size of the desired stimulus so the Stimulator will crash if the stimulus is too close to the edge of the screen. This was done to move the stimulus in a line or ellipse.

This is not a problem with the manual mapper because that uses an off screen buffer the size of the full screen instead.